Move CorsPolicyProvider to Jellyfin.Server.Configuration
